crawler/requirements.txt
twscrape
PyYAML
openai
aiofiles